OBGM-47 Move bind Domain methods to ImportDataService
